// src/interfaces/meeting_attendees.interface.ts
export interface MeetingAttendees {
    attendee_id: string;
    meeting_id: string;    
    staff_id: string;
    first_name: string;
    last_name: string;
    sub_phase_flag: number;
    sub_phase_activity: string;
    check_in_time: string;
    check_out_time: string;
    duration: string;
    imei: string;
    app_version: string;
    hub_id: string;
  }
  